# Fix Unmarshaling of relationships with only links