Fixes #34

Relationships without a `"data"` member will not successfully unmarshal, so long as at least one other of the required members exists.

**NOTE**: Even if unmarshaling is successful, we don't necessarily do anything with the other fields (like link objects or metas in relationships). If deemed important, this can be tackled in a follow-up PR.

Benchmarks before the change:
```
❯ go test -bench '^BenchmarkUnmarshal'
goos: darwin
goarch: arm64
pkg: github.com/DataDog/jsonapi
BenchmarkUnmarshal/ArticleSimple-10                               277360              4323 ns/op
BenchmarkUnmarshal/ArticleSimpleWithToplevelMeta-10               213162              5480 ns/op
BenchmarkUnmarshal/ArticleComplex-10                               42676             28024 ns/op
BenchmarkUnmarshal/ArticleComplexDisableNameValidation-10          51916             22794 ns/op
BenchmarkUnmarshal/ArticlesComplex-10                               8203            141519 ns/op
BenchmarkUnmarshal/ArticlesComplexDisableNameValidation-10         10000            116609 ns/op
PASS
ok      github.com/DataDog/jsonapi      8.842s
```

Benchmarks after the change

```
❯ go test -bench '^BenchmarkUnmarshal'
goos: darwin
goarch: arm64
pkg: github.com/DataDog/jsonapi
BenchmarkUnmarshal/ArticleSimple-10                               206233              5052 ns/op
BenchmarkUnmarshal/ArticleSimpleWithToplevelMeta-10               193520              6204 ns/op
BenchmarkUnmarshal/ArticleComplex-10                               33628             35616 ns/op
BenchmarkUnmarshal/ArticleComplexDisableNameValidation-10          39631             30107 ns/op
BenchmarkUnmarshal/ArticlesComplex-10                               6302            181894 ns/op
BenchmarkUnmarshal/ArticlesComplexDisableNameValidation-10          7180            158011 ns/op
PASS
ok      github.com/DataDog/jsonapi      7.898s
```

For a large complex document, this change results in approximately 0.05 additional milliseconds spent Unmarshaling.